refactor: continue handler decomposition and tighten analyze input validation
