+++ acestep/core/generation/handler/__init__.py
@@ -1,8 +1,29 @@
 """Handler decomposition components."""
 
+from .audio_codes import AudioCodesMixin
+from .batch_prep import BatchPrepMixin
 from .diffusion import DiffusionMixin
 from .init_service import InitServiceMixin
+from .io_audio import IoAudioMixin
 from .lora_manager import LoraManagerMixin
+from .memory_utils import MemoryUtilsMixin
+from .metadata_utils import MetadataMixin
+from .padding_utils import PaddingMixin
+from .prompt_utils import PromptMixin
 from .progress import ProgressMixin
+from .task_utils import TaskUtilsMixin
 
-__all__ = ["DiffusionMixin", "InitServiceMixin", "LoraManagerMixin", "ProgressMixin"]
+__all__ = [
+    "AudioCodesMixin",
+    "BatchPrepMixin",
+    "DiffusionMixin",
+    "InitServiceMixin",
+    "IoAudioMixin",
+    "LoraManagerMixin",
+    "MemoryUtilsMixin",
+    "MetadataMixin",
+    "PaddingMixin",
+    "PromptMixin",
+    "ProgressMixin",
+    "TaskUtilsMixin",
+]
