+++ acestep/gradio_ui/events/generation_handlers.py
@@ -7,6 +7,7 @@ import sys
 import json
 import random
 import glob
+import re
 import gradio as gr
 from typing import Optional, List, Tuple
 from loguru import logger
@@ -761,6 +762,13 @@ def convert_src_audio_to_codes_wrapper(dit_handler, src_audio):
     return codes_string
 
 
+def _contains_audio_code_tokens(codes_string: str) -> bool:
+    """Return True when a string contains at least one serialized audio-code token."""
+    if not isinstance(codes_string, str):
+        return False
+    return bool(re.search(r"<\|audio_code_\d+\|>", codes_string))
+
+
 def analyze_src_audio(dit_handler, llm_handler, src_audio, constrained_decoding_debug=False):
     """Analyze source audio: convert to codes, then transcribe to caption/lyrics/metas.
 
@@ -791,6 +799,10 @@ def analyze_src_audio(dit_handler, llm_handler, src_audio, constrained_decoding_
 
     if not codes_string or not codes_string.strip():
         gr.Warning("Audio conversion produced empty codes.")
+        return _err
+
+    if not _contains_audio_code_tokens(codes_string):
+        gr.Warning("Source file is not valid audio or conversion failed (no audio codes detected).")
         return _err
 
     # Step 2: Transcribe codes to caption/lyrics/metas via LLM
