+++ acestep/handler.py
@@ -38,16 +38,20 @@ from acestep.model_downloader import (
     check_model_exists,
     get_checkpoints_dir,
 )
-from acestep.constants import (
-    TASK_INSTRUCTIONS,
-    SFT_GEN_PROMPT,
-    DEFAULT_DIT_INSTRUCTION,
-)
+from acestep.constants import DEFAULT_DIT_INSTRUCTION, SFT_GEN_PROMPT, TASK_INSTRUCTIONS
 from acestep.core.generation.handler import (
+    AudioCodesMixin,
+    BatchPrepMixin,
     DiffusionMixin,
     InitServiceMixin,
+    IoAudioMixin,
     LoraManagerMixin,
+    MemoryUtilsMixin,
+    MetadataMixin,
+    PaddingMixin,
     ProgressMixin,
+    PromptMixin,
+    TaskUtilsMixin,
 )
 from acestep.dit_alignment_score import MusicStampsAligner, MusicLyricScorer
 from acestep.gpu_config import get_gpu_memory_gb, get_global_gpu_config, get_effective_free_vram_gb
@@ -56,7 +60,20 @@ from acestep.gpu_config import get_gpu_memory_gb, get_global_gpu_config, get_eff
 warnings.filterwarnings("ignore")
 
 
-class AceStepHandler(DiffusionMixin, InitServiceMixin, LoraManagerMixin, ProgressMixin):
+class AceStepHandler(
+    DiffusionMixin,
+    AudioCodesMixin,
+    BatchPrepMixin,
+    IoAudioMixin,
+    InitServiceMixin,
+    LoraManagerMixin,
+    MemoryUtilsMixin,
+    MetadataMixin,
+    PaddingMixin,
+    ProgressMixin,
+    PromptMixin,
+    TaskUtilsMixin,
+):
     """ACE-Step Business Logic Handler"""
     
     def __init__(self):
@@ -856,1102 +873,6 @@ class AceStepHandler(DiffusionMixin, InitServiceMixin, LoraManagerMixin, Progres
             return f"Switched to training preset (quantization disabled).\n{status}", True
         return f"Failed to switch to training preset.\n{status}", False
     
-    def process_target_audio(self, audio_file) -> Optional[torch.Tensor]:
-        """Process target audio"""
-        if audio_file is None:
-            return None
-        
-        try:
-            # Load audio using soundfile
-            audio_np, sr = sf.read(audio_file, dtype='float32')
-            # Convert to torch: [samples, channels] or [samples] -> [channels, samples]
-            if audio_np.ndim == 1:
-                audio = torch.from_numpy(audio_np).unsqueeze(0)
-            else:
-                audio = torch.from_numpy(audio_np.T)
-            
-            # Normalize to stereo 48kHz
-            audio = self._normalize_audio_to_stereo_48k(audio, sr)
-            
-            return audio
-        except Exception as e:
-            logger.exception("[process_target_audio] Error processing target audio")
-            return None
-    
-    def _parse_audio_code_string(self, code_str: str) -> List[int]:
-        """Extract integer audio codes from prompt tokens like <|audio_code_123|>.
-        Code values are clamped to valid range [0, 63999] (codebook size = 64000).
-        """
-        if not code_str:
-            return []
-        try:
-            MAX_AUDIO_CODE = 63999  # Maximum valid audio code value (codebook size = 64000)
-            codes = []
-            clamped_count = 0
-            for x in re.findall(r"<\|audio_code_(\d+)\|>", code_str):
-                code_value = int(x)
-                # Clamp code value to valid range [0, MAX_AUDIO_CODE]
-                clamped_value = max(0, min(code_value, MAX_AUDIO_CODE))
-                if clamped_value != code_value:
-                    clamped_count += 1
-                    logger.warning(f"[_parse_audio_code_string] Clamped audio code value from {code_value} to {clamped_value}")
-                codes.append(clamped_value)
-            if clamped_count > 0:
-                logger.warning(f"[_parse_audio_code_string] Clamped {clamped_count} audio code value(s) to valid range [0, {MAX_AUDIO_CODE}]")
-            return codes
-        except Exception as e:
-            logger.debug(f"[_parse_audio_code_string] Failed to parse audio code string: {e}")
-            return []
-    
-    def _decode_audio_codes_to_latents(self, code_str: str) -> Optional[torch.Tensor]:
-        """
-        Convert serialized audio code string into 25Hz latents using model quantizer/detokenizer.
-        
-        Note: Code values are already clamped to valid range [0, 63999] by _parse_audio_code_string(),
-        ensuring indices are within the quantizer's codebook size (64000).
-        """
-        if self.model is None or not hasattr(self.model, 'tokenizer') or not hasattr(self.model, 'detokenizer'):
-            return None
-        
-        code_ids = self._parse_audio_code_string(code_str)
-        if len(code_ids) == 0:
-            return None
-        
-        with self._load_model_context("model"):
-            quantizer = self.model.tokenizer.quantizer
-            detokenizer = self.model.detokenizer
-            
-            num_quantizers = getattr(quantizer, "num_quantizers", 1)
-            # Create indices tensor: [T_5Hz]
-            # Note: code_ids are already clamped to [0, 63999] by _parse_audio_code_string()
-            indices = torch.tensor(code_ids, device=self.device, dtype=torch.long)  # [T_5Hz]
-            
-            indices = indices.unsqueeze(0).unsqueeze(-1)  # [1, T_5Hz, 1]
-            
-            # Get quantized representation from indices
-            # The quantizer expects [batch, T_5Hz] format and handles quantizer dimension internally
-            quantized = quantizer.get_output_from_indices(indices)
-            if quantized.dtype != self.dtype:
-                quantized = quantized.to(self.dtype)
-            
-            # Detokenize to 25Hz: [1, T_5Hz, dim] -> [1, T_25Hz, dim]
-            lm_hints_25hz = detokenizer(quantized)
-            return lm_hints_25hz
-    
-    def _create_default_meta(self) -> str:
-        """Create default metadata string."""
-        return (
-            "- bpm: N/A\n"
-            "- timesignature: N/A\n" 
-            "- keyscale: N/A\n"
-            "- duration: 30 seconds\n"
-        )
-    
-    def _dict_to_meta_string(self, meta_dict: Dict[str, Any]) -> str:
-        """Convert metadata dict to formatted string."""
-        bpm = meta_dict.get('bpm', meta_dict.get('tempo', 'N/A'))
-        timesignature = meta_dict.get('timesignature', meta_dict.get('time_signature', 'N/A'))
-        keyscale = meta_dict.get('keyscale', meta_dict.get('key', meta_dict.get('scale', 'N/A')))
-        duration = meta_dict.get('duration', meta_dict.get('length', 30))
-
-        # Format duration
-        if isinstance(duration, (int, float)):
-            duration = f"{int(duration)} seconds"
-        elif not isinstance(duration, str):
-            duration = "30 seconds"
-        
-        return (
-            f"- bpm: {bpm}\n"
-            f"- timesignature: {timesignature}\n"
-            f"- keyscale: {keyscale}\n"
-            f"- duration: {duration}\n"
-        )
-    
-    def _parse_metas(self, metas: List[Union[str, Dict[str, Any]]]) -> List[str]:
-        """
-        Parse and normalize metadata with fallbacks.
-        
-        Args:
-            metas: List of metadata (can be strings, dicts, or None)
-            
-        Returns:
-            List of formatted metadata strings
-        """
-        parsed_metas = []
-        for meta in metas:
-            if meta is None:
-                # Default fallback metadata
-                parsed_meta = self._create_default_meta()
-            elif isinstance(meta, str):
-                # Already formatted string
-                parsed_meta = meta
-            elif isinstance(meta, dict):
-                # Convert dict to formatted string
-                parsed_meta = self._dict_to_meta_string(meta)
-            else:
-                # Fallback for any other type
-                parsed_meta = self._create_default_meta()
-            
-            parsed_metas.append(parsed_meta)
-        
-        return parsed_metas
-    
-    def build_dit_inputs(
-        self,
-        task: str,
-        instruction: Optional[str],
-        caption: str,
-        lyrics: str,
-        metas: Optional[Union[str, Dict[str, Any]]] = None,
-        vocal_language: str = "en",
-    ) -> Tuple[str, str]:
-        """
-        Build text inputs for the caption and lyric branches used by DiT.
-
-        Args:
-            task: Task name (e.g., text2music, cover, repaint); kept for logging/future branching.
-            instruction: Instruction text; default fallback matches service_generate behavior.
-            caption: Caption string (fallback if not in metas).
-            lyrics: Lyrics string.
-            metas: Metadata (str or dict); follows _parse_metas formatting.
-                   May contain 'caption' and 'language' fields from LM CoT output.
-            vocal_language: Language code for lyrics section (fallback if not in metas).
-
-        Returns:
-            (caption_input_text, lyrics_input_text)
-
-        Example:
-            caption_input, lyrics_input = handler.build_dit_inputs(
-                task="text2music",
-                instruction=None,
-                caption="A calm piano melody",
-                lyrics="la la la",
-                metas={"bpm": 90, "duration": 45, "caption": "LM generated caption", "language": "en"},
-                vocal_language="en",
-            )
-        """
-        # Align instruction formatting with _prepare_batch
-        final_instruction = self._format_instruction(instruction or DEFAULT_DIT_INSTRUCTION)
-
-        # Extract caption and language from metas if available (from LM CoT output)
-        # Fallback to user-provided values if not in metas
-        actual_caption = caption
-        actual_language = vocal_language
-        
-        if metas is not None:
-            # Parse metas to dict if it's a string
-            if isinstance(metas, str):
-                # Try to parse as dict-like string or use as-is
-                parsed_metas = self._parse_metas([metas])
-                if parsed_metas and isinstance(parsed_metas[0], dict):
-                    meta_dict = parsed_metas[0]
-                else:
-                    meta_dict = {}
-            elif isinstance(metas, dict):
-                meta_dict = metas
-            else:
-                meta_dict = {}
-            
-            # Extract caption from metas if available
-            if 'caption' in meta_dict and meta_dict['caption']:
-                actual_caption = str(meta_dict['caption'])
-            
-            # Extract language from metas if available
-            if 'language' in meta_dict and meta_dict['language']:
-                actual_language = str(meta_dict['language'])
-
-        parsed_meta = self._parse_metas([metas])[0]
-        caption_input = SFT_GEN_PROMPT.format(final_instruction, actual_caption, parsed_meta)
-        lyrics_input = self._format_lyrics(lyrics, actual_language)
-        return caption_input, lyrics_input
-    
-    def _get_text_hidden_states(self, text_prompt: str) -> Tuple[torch.Tensor, torch.Tensor]:
-        """Get text hidden states from text encoder."""
-        if self.text_tokenizer is None or self.text_encoder is None:
-            raise ValueError("Text encoder not initialized")
-        
-        with self._load_model_context("text_encoder"):
-            # Tokenize
-            text_inputs = self.text_tokenizer(
-                text_prompt,
-                padding="longest",
-                truncation=True,
-                max_length=256,
-                return_tensors="pt",
-            )
-            text_input_ids = text_inputs.input_ids.to(self.device)
-            text_attention_mask = text_inputs.attention_mask.to(self.device).bool()
-            
-            # Encode
-            with torch.inference_mode():
-                text_outputs = self.text_encoder(text_input_ids)
-                if hasattr(text_outputs, 'last_hidden_state'):
-                    text_hidden_states = text_outputs.last_hidden_state
-                elif isinstance(text_outputs, tuple):
-                    text_hidden_states = text_outputs[0]
-                else:
-                    text_hidden_states = text_outputs
-            
-            text_hidden_states = text_hidden_states.to(self.dtype)
-            
-            return text_hidden_states, text_attention_mask
-    
-    def extract_caption_from_sft_format(self, caption: str) -> str:
-        try:
-            if "# Instruction" in caption and "# Caption" in caption:
-                pattern = r'#\s*Caption\s*\n(.*?)(?:\n\s*#\s*Metas|$)'
-                match = re.search(pattern, caption, re.DOTALL)
-                if match:
-                    return match.group(1).strip()
-            return caption
-        except Exception as e:
-            logger.exception("[extract_caption_from_sft_format] Error extracting caption")
-            return caption
-
-    def prepare_seeds(self, actual_batch_size, seed, use_random_seed):
-        actual_seed_list: List[int] = []
-        seed_value_for_ui = ""
-
-        if use_random_seed:
-            # Generate brand new seeds and expose them back to the UI
-            actual_seed_list = [random.randint(0, 2 ** 32 - 1) for _ in range(actual_batch_size)]
-            seed_value_for_ui = ", ".join(str(s) for s in actual_seed_list)
-        else:
-            # Parse seed input: can be a single number, comma-separated numbers, or -1
-            # If seed is a string, try to parse it as comma-separated values
-            seed_list = []
-            if isinstance(seed, str):
-                # Handle string input (e.g., "123,456" or "-1")
-                seed_str_list = [s.strip() for s in seed.split(",")]
-                for s in seed_str_list:
-                    if s == "-1" or s == "":
-                        seed_list.append(-1)
-                    else:
-                        try:
-                            seed_list.append(int(float(s)))
-                        except (ValueError, TypeError) as e:
-                            logger.debug(f"[prepare_seeds] Failed to parse seed value '{s}': {e}")
-                            seed_list.append(-1)
-            elif seed is None or (isinstance(seed, (int, float)) and seed < 0):
-                # If seed is None or negative, use -1 for all items
-                seed_list = [-1] * actual_batch_size
-            elif isinstance(seed, (int, float)):
-                # Single seed value
-                seed_list = [int(seed)]
-            else:
-                # Fallback: use -1
-                seed_list = [-1] * actual_batch_size
-
-            # Process seed list according to rules:
-            # 1. If all are -1, generate different random seeds for each batch item
-            # 2. If one non-negative seed is provided and batch_size > 1, first uses that seed, rest are random
-            # 3. If more seeds than batch_size, use first batch_size seeds
-            # Check if user provided only one non-negative seed (not -1)
-            has_single_non_negative_seed = (len(seed_list) == 1 and seed_list[0] != -1)
-
-            for i in range(actual_batch_size):
-                if i < len(seed_list):
-                    seed_val = seed_list[i]
-                else:
-                    # If not enough seeds provided, use -1 (will generate random)
-                    seed_val = -1
-
-                # Special case: if only one non-negative seed was provided and batch_size > 1,
-                # only the first item uses that seed, others are random
-                if has_single_non_negative_seed and actual_batch_size > 1 and i > 0:
-                    # Generate random seed for remaining items
-                    actual_seed_list.append(random.randint(0, 2 ** 32 - 1))
-                elif seed_val == -1:
-                    # Generate a random seed for this item
-                    actual_seed_list.append(random.randint(0, 2 ** 32 - 1))
-                else:
-                    actual_seed_list.append(int(seed_val))
-
-            seed_value_for_ui = ", ".join(str(s) for s in actual_seed_list)
-        return actual_seed_list, seed_value_for_ui
-    
-    def prepare_metadata(self, bpm, key_scale, time_signature):
-        """Build metadata dict - use "N/A" as default for empty fields."""
-        return self._build_metadata_dict(bpm, key_scale, time_signature)
-    
-    def is_silence(self, audio):
-        return torch.all(audio.abs() < 1e-6)
-    
-    def _get_system_memory_gb(self) -> Optional[float]:
-        """Return total system RAM in GB when available."""
-        try:
-            page_size = os.sysconf("SC_PAGE_SIZE")
-            page_count = os.sysconf("SC_PHYS_PAGES")
-            if page_size and page_count:
-                return (page_size * page_count) / (1024 ** 3)
-        except (ValueError, OSError, AttributeError):
-            return None
-        return None
-
-    def _get_effective_mps_memory_gb(self) -> Optional[float]:
-        """Best-effort MPS memory estimate (recommended max or system RAM)."""
-        if hasattr(torch, "mps") and hasattr(torch.mps, "recommended_max_memory"):
-            try:
-                return torch.mps.recommended_max_memory() / (1024 ** 3)
-            except Exception:
-                pass
-        system_gb = self._get_system_memory_gb()
-        if system_gb is None:
-            return None
-        # Align with gpu_config: MPS can use ~75% of unified memory for GPU workloads.
-        return system_gb * 0.75
-
-    # Maximum VAE decode chunk size.  Larger chunks are faster but the
-    # PyTorch caching allocator may *reserve* significantly more VRAM than
-    # the peak *allocated* amount.  Empirical measurements (bf16 VAE,
-    # ~10 GB baseline from DiT + LM):
-    #   chunk  peak_alloc  peak_reserved
-    #    512     11.9 GB     12.7 GB
-    #   1024     13.1 GB     15.0 GB   ← dangerously close to 16 GB
-    #   1536     14.4 GB     17.2 GB   ← exceeds 16 GB
-    # Capping at 512 keeps reserved VRAM safely under 16 GB on consumer
-    # GPUs while the speed difference vs 1024/1536 is negligible for
-    # tiled decode (a few hundred ms).
-    VAE_DECODE_MAX_CHUNK_SIZE = 512
-
-    def _get_auto_decode_chunk_size(self) -> int:
-        """Choose a conservative VAE decode chunk size based on available memory.
-        
-        For CUDA GPUs, uses actual free VRAM to determine chunk size.
-        For MPS, uses effective memory estimate.
-        Larger chunks are faster but use more VRAM; smaller chunks are safer.
-        The result is capped at ``VAE_DECODE_MAX_CHUNK_SIZE`` to prevent the
-        PyTorch caching allocator from over-reserving VRAM on consumer GPUs.
-        """
-        override = os.environ.get("ACESTEP_VAE_DECODE_CHUNK_SIZE")
-        if override:
-            try:
-                value = int(override)
-                if value > 0:
-                    return value  # explicit override bypasses the cap
-            except ValueError:
-                pass
-
-        max_chunk = self.VAE_DECODE_MAX_CHUNK_SIZE
-
-        if self.device == "mps":
-            mem_gb = self._get_effective_mps_memory_gb()
-            if mem_gb is not None:
-                if mem_gb >= 48:
-                    return min(1536, max_chunk)
-                if mem_gb >= 24:
-                    return min(1024, max_chunk)
-            return min(512, max_chunk)
-        
-        # CUDA: use effective free VRAM (respects per-process memory fraction) to pick chunk size
-        if self.device == "cuda" or (isinstance(self.device, str) and self.device.startswith("cuda")):
-            try:
-                free_gb = get_effective_free_vram_gb()
-            except Exception:
-                free_gb = 0
-            logger.debug(f"[_get_auto_decode_chunk_size] Effective free VRAM: {free_gb:.2f} GB")
-            # VAE decode peak VRAM (allocated) scales roughly with chunk_size.
-            # Empirical: chunk_size=512 needs ~1.3 GB, 1024 needs ~2.6 GB, 1536 needs ~3.9 GB
-            # chunk_size=128 needs ~0.3 GB, chunk_size=64 needs ~0.3 GB
-            if free_gb >= 8.0:
-                return min(512, max_chunk)
-            elif free_gb >= 5.0:
-                return min(512, max_chunk)
-            elif free_gb >= 2.5:
-                return min(512, max_chunk)
-            elif free_gb >= 1.0:
-                return 256
-            elif free_gb >= 0.5:
-                return 128  # Very tight VRAM
-            else:
-                return 64   # Extremely tight VRAM — minimal chunk
-        
-        return min(512, max_chunk)
-
-    def _should_offload_wav_to_cpu(self) -> bool:
-        """Decide whether to offload decoded wavs to CPU for memory safety.
-        
-        For CUDA GPUs with >=24 GB free, keep on GPU for speed.
-        For MPS with >=32 GB, keep on GPU.
-        Otherwise offload to CPU to avoid OOM during concatenation.
-        """
-        override = os.environ.get("ACESTEP_MPS_DECODE_OFFLOAD")
-        if override:
-            return override.lower() in ("1", "true", "yes")
-        if self.device == "mps":
-            mem_gb = self._get_effective_mps_memory_gb()
-            if mem_gb is not None and mem_gb >= 32:
-                return False
-            return True
-        # CUDA: offload unless plenty of free VRAM
-        if self.device == "cuda" or (isinstance(self.device, str) and self.device.startswith("cuda")):
-            try:
-                free_gb = get_effective_free_vram_gb()
-                logger.debug(f"[_should_offload_wav_to_cpu] Effective free VRAM: {free_gb:.2f} GB")
-                if free_gb >= 24.0:
-                    return False
-            except Exception:
-                pass
-        return True
-
-    def _vram_guard_reduce_batch(
-        self,
-        batch_size: int,
-        audio_duration: Optional[float] = None,
-        use_lm: bool = False,
-    ) -> int:
-        """Pre-inference VRAM guard: auto-reduce batch_size if free VRAM is tight.
-        
-        Rough activation estimate per batch element:
-          - DiT forward pass: ~0.8 GB per sample at 60s, scales linearly with duration
-          - LM inference: KV cache is pre-allocated so batch doesn't change it much
-          - VAE decode: handled separately via tiled_decode
-        
-        We leave a 1.5 GB safety margin for CUDA allocator fragmentation.
-        
-        IMPORTANT: When offload_to_cpu is True, the LM model (especially vllm
-        backend) may still be on GPU when this guard runs, but it will be
-        offloaded or its memory reclaimed before DiT actually needs the VRAM.
-        In that case we trust the static GPU tier config limits (which have been
-        empirically validated) and skip the dynamic VRAM check.
-        """
-        if batch_size <= 1:
-            return batch_size
-
-        device = self.device
-        if device == "cpu" or device == "mps":
-            return batch_size  # No CUDA VRAM to guard
-
-        # When CPU offload is enabled, the current free VRAM is misleading because
-        # the LM (vllm KV cache + weights) may still be on GPU at this point but
-        # will be released/reclaimed before DiT actually uses the VRAM.  The static
-        # GPU tier configs already encode safe batch limits that were empirically
-        # validated with offload enabled, so trust them.
-        #
-        # Use the more conservative max_batch_size_with_lm as the threshold since
-        # the handler doesn't know if LM was used upstream.  This is safe because
-        # max_batch_size_with_lm <= max_batch_size_without_lm for all tiers.
-        if self.offload_to_cpu:
-            gpu_config = get_global_gpu_config()
-            if gpu_config is not None:
-                tier_max = gpu_config.max_batch_size_with_lm
-                if batch_size <= tier_max:
-                    logger.debug(
-                        f"[VRAM guard] offload_to_cpu=True, batch_size={batch_size} <= "
-                        f"tier limit {tier_max} — skipping dynamic VRAM check "
-                        f"(LM will be offloaded before DiT runs)"
-                    )
-                    return batch_size
-                # batch_size exceeds tier limit — fall through to dynamic check
-
-        try:
-            free_gb = get_effective_free_vram_gb()
-        except Exception:
-            return batch_size
-
-        # Estimate per-sample activation cost for DiT
-        duration_sec = float(audio_duration) if audio_duration and float(audio_duration) > 0 else 60.0
-        # Empirical observation: DiT activation memory per extra batch element is
-        # relatively modest because the latent is processed in a single forward pass
-        # and flash-attention keeps peak memory low.  Measured values:
-        #   - 60s turbo, noLM, batch 4 → ~13.3 GB total on 16GB GPU
-        #     (model ~8.5 GB + 4 × ~0.8 GB activations ≈ 11.7 GB + overhead)
-        #   - 208s turbo, batch 1 → peak 9.3 GB (model ~8.9 GB + ~0.4 GB activation)
-        # The old formula (0.8 * duration/60) heavily overestimates for long durations
-        # because activation memory scales sub-linearly with latent length (flash attn).
-        # Use a more conservative formula: base 0.5 GB + 0.15 GB per 60s beyond 60s.
-        per_sample_gb = 0.5 + max(0.0, 0.15 * (duration_sec - 60.0) / 60.0)
-        # If using cfg (base model), double the per-sample cost
-        if hasattr(self, 'model') and self.model is not None:
-            model_name = getattr(self, 'config_path', '') or ''
-            if 'base' in model_name.lower():
-                per_sample_gb *= 2.0
-
-        safety_margin_gb = 1.5
-        available_for_batch = free_gb - safety_margin_gb
-
-        if available_for_batch <= 0:
-            logger.warning(
-                f"[VRAM guard] Only {free_gb:.1f} GB free — reducing batch_size to 1"
-            )
-            return 1
-
-        max_safe_batch = max(1, int(available_for_batch / per_sample_gb))
-        if max_safe_batch < batch_size:
-            logger.warning(
-                f"[VRAM guard] Free VRAM {free_gb:.1f} GB can safely fit ~{max_safe_batch} samples "
-                f"(requested {batch_size}). Reducing batch_size to {max_safe_batch}."
-            )
-            return max_safe_batch
-
-        return batch_size
-    def _get_vae_dtype(self, device: Optional[str] = None) -> torch.dtype:
-        """Get VAE dtype based on target device and GPU tier."""
-        target_device = device or self.device
-        if target_device in ["cuda", "xpu"]:
-            return torch.bfloat16
-        if target_device == "mps":
-            return torch.float16
-        if target_device == "cpu":
-            # CPU float16/bfloat16 VAE paths are typically much slower and less stable.
-            return torch.float32
-        return self.dtype
-    
-    def _format_instruction(self, instruction: str) -> str:
-        """Format instruction to ensure it ends with colon."""
-        if not instruction.endswith(":"):
-            instruction = instruction + ":"
-        return instruction
-    
-    def _normalize_audio_to_stereo_48k(self, audio: torch.Tensor, sr: int) -> torch.Tensor:
-        """
-        Normalize audio to stereo 48kHz format.
-        
-        Args:
-            audio: Audio tensor [channels, samples] or [samples]
-            sr: Sample rate
-            
-        Returns:
-            Normalized audio tensor [2, samples] at 48kHz
-        """
-        # Convert to stereo (duplicate channel if mono)
-        if audio.shape[0] == 1:
-            audio = torch.cat([audio, audio], dim=0)
-        
-        # Keep only first 2 channels
-        audio = audio[:2]
-        
-        # Resample to 48kHz if needed
-        if sr != 48000:
-            audio = torchaudio.transforms.Resample(sr, 48000)(audio)
-        
-        # Clamp values to [-1.0, 1.0]
-        audio = torch.clamp(audio, -1.0, 1.0)
-        
-        return audio
-    
-    def _normalize_audio_code_hints(self, audio_code_hints: Optional[Union[str, List[str]]], batch_size: int) -> List[Optional[str]]:
-        """Normalize audio_code_hints to list of correct length."""
-        if audio_code_hints is None:
-            normalized = [None] * batch_size
-        elif isinstance(audio_code_hints, str):
-            normalized = [audio_code_hints] * batch_size
-        elif len(audio_code_hints) == 1 and batch_size > 1:
-            normalized = audio_code_hints * batch_size
-        elif len(audio_code_hints) != batch_size:
-            # Pad or truncate to match batch_size
-            normalized = list(audio_code_hints[:batch_size])
-            while len(normalized) < batch_size:
-                normalized.append(None)
-        else:
-            normalized = list(audio_code_hints)
-        
-        # Clean up: convert empty strings to None
-        normalized = [hint if isinstance(hint, str) and hint.strip() else None for hint in normalized]
-        return normalized
-    
-    def _normalize_instructions(self, instructions: Optional[Union[str, List[str]]], batch_size: int, default: Optional[str] = None) -> List[str]:
-        """Normalize instructions to list of correct length."""
-        if instructions is None:
-            default_instruction = default or DEFAULT_DIT_INSTRUCTION
-            return [default_instruction] * batch_size
-        elif isinstance(instructions, str):
-            return [instructions] * batch_size
-        elif len(instructions) == 1:
-            return instructions * batch_size
-        elif len(instructions) != batch_size:
-            # Pad or truncate to match batch_size
-            normalized = list(instructions[:batch_size])
-            default_instruction = default or DEFAULT_DIT_INSTRUCTION
-            while len(normalized) < batch_size:
-                normalized.append(default_instruction)
-            return normalized
-        else:
-            return list(instructions)
-    
-    def _format_lyrics(self, lyrics: str, language: str) -> str:
-        """Format lyrics text with language header."""
-        return f"# Languages\n{language}\n\n# Lyric\n{lyrics}<|endoftext|>"
-    
-    def _pad_sequences(self, sequences: List[torch.Tensor], max_length: int, pad_value: int = 0) -> torch.Tensor:
-        """Pad sequences to same length."""
-        return torch.stack([
-            torch.nn.functional.pad(seq, (0, max_length - len(seq)), 'constant', pad_value)
-            for seq in sequences
-        ])
-    
-    def _extract_caption_and_language(self, metas: List[Union[str, Dict[str, Any]]], captions: List[str], vocal_languages: List[str]) -> Tuple[List[str], List[str]]:
-        """Extract caption and language from metas with fallback to provided values."""
-        actual_captions = list(captions)
-        actual_languages = list(vocal_languages)
-        
-        for i, meta in enumerate(metas):
-            if i >= len(actual_captions):
-                break
-                
-            meta_dict = None
-            if isinstance(meta, str):
-                parsed = self._parse_metas([meta])
-                if parsed and isinstance(parsed[0], dict):
-                    meta_dict = parsed[0]
-            elif isinstance(meta, dict):
-                meta_dict = meta
-            
-            if meta_dict:
-                if 'caption' in meta_dict and meta_dict['caption']:
-                    actual_captions[i] = str(meta_dict['caption'])
-                if 'language' in meta_dict and meta_dict['language']:
-                    actual_languages[i] = str(meta_dict['language'])
-        
-        return actual_captions, actual_languages
-    
-    def _encode_audio_to_latents(self, audio: torch.Tensor) -> torch.Tensor:
-        """
-        Encode audio to latents using VAE with tiled encoding for long audio.
-        
-        Args:
-            audio: Audio tensor [channels, samples] or [batch, channels, samples]
-            
-        Returns:
-            Latents tensor [T, D] or [batch, T, D]
-        """
-        # Save original dimension info BEFORE modifying audio
-        input_was_2d = (audio.dim() == 2)
-        
-        # Ensure batch dimension
-        if input_was_2d:
-            audio = audio.unsqueeze(0)
-        
-        # Use tiled_encode for memory-efficient encoding
-        # tiled_encode handles device transfer and dtype conversion internally
-        with torch.inference_mode():
-            latents = self.tiled_encode(audio, offload_latent_to_cpu=True)
-        
-        # Move back to device and cast to model dtype
-        latents = latents.to(self.device).to(self.dtype)
-        
-        # Transpose: [batch, d, T] -> [batch, T, d]
-        latents = latents.transpose(1, 2)
-        
-        # Remove batch dimension if input didn't have it
-        if input_was_2d:
-            latents = latents.squeeze(0)
-        
-        return latents
-    
-    def _build_metadata_dict(self, bpm: Optional[Union[int, str]], key_scale: str, time_signature: str, duration: Optional[float] = None) -> Dict[str, Any]:
-        """
-        Build metadata dictionary with default values.
-        
-        Args:
-            bpm: BPM value (optional)
-            key_scale: Key/scale string
-            time_signature: Time signature string
-            duration: Duration in seconds (optional)
-            
-        Returns:
-            Metadata dictionary
-        """
-        metadata_dict = {}
-        if bpm:
-            metadata_dict["bpm"] = bpm
-        else:
-            metadata_dict["bpm"] = "N/A"
-
-        if key_scale.strip():
-            metadata_dict["keyscale"] = key_scale
-        else:
-            metadata_dict["keyscale"] = "N/A"
-
-        if time_signature.strip() and time_signature != "N/A" and time_signature:
-            metadata_dict["timesignature"] = time_signature
-        else:
-            metadata_dict["timesignature"] = "N/A"
-        
-        # Add duration if provided
-        if duration is not None:
-            metadata_dict["duration"] = f"{int(duration)} seconds"
-        
-        return metadata_dict
-
-    def generate_instruction(
-        self,
-        task_type: str,
-        track_name: Optional[str] = None,
-        complete_track_classes: Optional[List[str]] = None
-    ) -> str:
-        if task_type == "text2music":
-            return TASK_INSTRUCTIONS["text2music"]
-        elif task_type == "repaint":
-            return TASK_INSTRUCTIONS["repaint"]
-        elif task_type == "cover":
-            return TASK_INSTRUCTIONS["cover"]
-        elif task_type == "extract":
-            if track_name:
-                # Convert to uppercase
-                track_name_upper = track_name.upper()
-                return TASK_INSTRUCTIONS["extract"].format(TRACK_NAME=track_name_upper)
-            else:
-                return TASK_INSTRUCTIONS["extract_default"]
-        elif task_type == "lego":
-            if track_name:
-                # Convert to uppercase
-                track_name_upper = track_name.upper()
-                return TASK_INSTRUCTIONS["lego"].format(TRACK_NAME=track_name_upper)
-            else:
-                return TASK_INSTRUCTIONS["lego_default"]
-        elif task_type == "complete":
-            if complete_track_classes and len(complete_track_classes) > 0:
-                # Convert to uppercase and join with " | "
-                track_classes_upper = [t.upper() for t in complete_track_classes]
-                complete_track_classes_str = " | ".join(track_classes_upper)
-                return TASK_INSTRUCTIONS["complete"].format(TRACK_CLASSES=complete_track_classes_str)
-            else:
-                return TASK_INSTRUCTIONS["complete_default"]
-        else:
-            return TASK_INSTRUCTIONS["text2music"]
-    
-    def process_reference_audio(self, audio_file) -> Optional[torch.Tensor]:
-        if audio_file is None:
-            return None
-            
-        try:
-            # Load audio file
-            audio, sr = torchaudio.load(audio_file)
-            
-            logger.debug(f"[process_reference_audio] Reference audio shape: {audio.shape}")
-            logger.debug(f"[process_reference_audio] Reference audio sample rate: {sr}")
-            logger.debug(f"[process_reference_audio] Reference audio duration: {audio.shape[-1] / 48000.0} seconds")
-            
-            # Normalize to stereo 48kHz
-            audio = self._normalize_audio_to_stereo_48k(audio, sr)
-            
-            is_silence = self.is_silence(audio)
-            if is_silence:
-                return None
-            
-            # Target length: 30 seconds at 48kHz
-            target_frames = 30 * 48000
-            segment_frames = 10 * 48000  # 10 seconds per segment
-            
-            # If audio is less than 30 seconds, repeat to at least 30 seconds
-            if audio.shape[-1] < target_frames:
-                repeat_times = math.ceil(target_frames / audio.shape[-1])
-                audio = audio.repeat(1, repeat_times)
-            # If audio is greater than or equal to 30 seconds, no operation needed
-            
-            # For all cases, select random 10-second segments from front, middle, and back
-            # then concatenate them to form 30 seconds
-            total_frames = audio.shape[-1]
-            segment_size = total_frames // 3
-            
-            # Front segment: [0, segment_size]
-            front_start = random.randint(0, max(0, segment_size - segment_frames))
-            front_audio = audio[:, front_start:front_start + segment_frames]
-            
-            # Middle segment: [segment_size, 2*segment_size]
-            middle_start = segment_size + random.randint(0, max(0, segment_size - segment_frames))
-            middle_audio = audio[:, middle_start:middle_start + segment_frames]
-            
-            # Back segment: [2*segment_size, total_frames]
-            back_start = 2 * segment_size + random.randint(0, max(0, (total_frames - 2 * segment_size) - segment_frames))
-            back_audio = audio[:, back_start:back_start + segment_frames]
-            
-            # Concatenate three segments to form 30 seconds
-            audio = torch.cat([front_audio, middle_audio, back_audio], dim=-1)
-            
-            return audio
-            
-        except Exception as e:
-            logger.exception("[process_reference_audio] Error processing reference audio")
-            return None
-
-    def process_src_audio(self, audio_file) -> Optional[torch.Tensor]:
-        if audio_file is None:
-            return None
-            
-        try:
-            # Load audio file
-            audio, sr = torchaudio.load(audio_file)
-            
-            # Normalize to stereo 48kHz
-            audio = self._normalize_audio_to_stereo_48k(audio, sr)
-            
-            return audio
-            
-        except Exception as e:
-            logger.exception("[process_src_audio] Error processing source audio")
-            return None
-    
-    def convert_src_audio_to_codes(self, audio_file) -> str:
-        """
-        Convert uploaded source audio to audio codes string.
-        
-        Args:
-            audio_file: Path to audio file or None
-            
-        Returns:
-            Formatted codes string like '<|audio_code_123|><|audio_code_456|>...' or error message
-        """
-        if audio_file is None:
-            return "❌ Please upload source audio first"
-        
-        if self.model is None or self.vae is None:
-            return "❌ Model not initialized. Please initialize the service first."
-        
-        try:
-            # Process audio file
-            processed_audio = self.process_src_audio(audio_file)
-            if processed_audio is None:
-                return "❌ Failed to process audio file"
-            
-            # Encode audio to latents using VAE
-            with torch.inference_mode():
-                with self._load_model_context("vae"):
-                    # Check if audio is silence
-                    if self.is_silence(processed_audio.unsqueeze(0)):
-                        return "❌ Audio file appears to be silent"
-                    
-                    # Encode to latents using helper method
-                    latents = self._encode_audio_to_latents(processed_audio)  # [T, d]
-                
-                # Create attention mask for latents
-                attention_mask = torch.ones(latents.shape[0], dtype=torch.bool, device=self.device)
-                
-                # Tokenize latents to get code indices
-                with self._load_model_context("model"):
-                    # Prepare latents for tokenize: [T, d] -> [1, T, d]
-                    hidden_states = latents.unsqueeze(0)  # [1, T, d]
-                    
-                    # Call tokenize method
-                    # tokenize returns: (quantized, indices, attention_mask)
-                    _, indices, _ = self.model.tokenize(hidden_states, self.silence_latent, attention_mask.unsqueeze(0))
-                    
-                    # Format indices as code string
-                    # indices shape: [1, T_5Hz] or [1, T_5Hz, num_quantizers]
-                    # Flatten and convert to list
-                    indices_flat = indices.flatten().cpu().tolist()
-                    codes_string = "".join([f"<|audio_code_{idx}|>" for idx in indices_flat])
-                    
-                    logger.info(f"[convert_src_audio_to_codes] Generated {len(indices_flat)} audio codes")
-                    return codes_string
-                    
-        except Exception as e:
-            error_msg = f"❌ Error converting audio to codes: {str(e)}\n{traceback.format_exc()}"
-            logger.exception("[convert_src_audio_to_codes] Error converting audio to codes")
-            return error_msg
-        
-    def prepare_batch_data(
-        self,
-        actual_batch_size,
-        processed_src_audio,
-        audio_duration,
-        captions,
-        lyrics,
-        vocal_language,
-        instruction,
-        bpm,
-        key_scale,
-        time_signature
-    ):
-        pure_caption = self.extract_caption_from_sft_format(captions)
-        captions_batch = [pure_caption] * actual_batch_size
-        instructions_batch = [instruction] * actual_batch_size
-        lyrics_batch = [lyrics] * actual_batch_size
-        vocal_languages_batch = [vocal_language] * actual_batch_size
-        # Calculate duration for metadata
-        calculated_duration = None
-        if processed_src_audio is not None:
-            calculated_duration = processed_src_audio.shape[-1] / 48000.0
-        elif audio_duration is not None and float(audio_duration) > 0:
-            calculated_duration = float(audio_duration)
-
-        # Build metadata dict - use "N/A" as default for empty fields
-        metadata_dict = self._build_metadata_dict(bpm, key_scale, time_signature, calculated_duration)
-
-        # Format metadata - inference service accepts dict and will convert to string
-        # Create a copy for each batch item (in case we modify it)
-        metas_batch = [metadata_dict.copy() for _ in range(actual_batch_size)]
-        return captions_batch, instructions_batch, lyrics_batch, vocal_languages_batch, metas_batch
-    
-    def determine_task_type(self, task_type, audio_code_string):
-        # Determine task type - repaint and lego tasks can have repainting parameters
-        # Other tasks (cover, text2music, extract, complete) should NOT have repainting
-        is_repaint_task = (task_type == "repaint")
-        is_lego_task = (task_type == "lego")
-        is_cover_task = (task_type == "cover")
-
-        has_codes = False
-        if isinstance(audio_code_string, list):
-            has_codes = any((c or "").strip() for c in audio_code_string)
-        else:
-            has_codes = bool(audio_code_string and str(audio_code_string).strip())
-
-        if has_codes:
-            is_cover_task = True
-        # Both repaint and lego tasks can use repainting parameters for chunk mask
-        can_use_repainting = is_repaint_task or is_lego_task
-        return is_repaint_task, is_lego_task, is_cover_task, can_use_repainting
-
-    def create_target_wavs(self, duration_seconds: float) -> torch.Tensor:
-        try:
-            # Ensure minimum precision of 100ms
-            duration_seconds = max(0.1, round(duration_seconds, 1))
-            # Calculate frames for 48kHz stereo
-            frames = int(duration_seconds * 48000)
-            # Create silent stereo audio
-            target_wavs = torch.zeros(2, frames)
-            return target_wavs
-        except Exception as e:
-            logger.exception("[create_target_wavs] Error creating target audio")
-            # Fallback to 30 seconds if error
-            return torch.zeros(2, 30 * 48000)
-    
-    def prepare_padding_info(
-        self,
-        actual_batch_size,
-        processed_src_audio,
-        audio_duration,
-        repainting_start,
-        repainting_end,
-        is_repaint_task,
-        is_lego_task,
-        is_cover_task,
-        can_use_repainting,
-    ):
-        target_wavs_batch = []
-        # Store padding info for each batch item to adjust repainting coordinates
-        padding_info_batch = []
-        for i in range(actual_batch_size):
-            if processed_src_audio is not None:
-                if is_cover_task:
-                    # Cover task: Use src_audio directly without padding
-                    batch_target_wavs = processed_src_audio
-                    padding_info_batch.append({
-                        'left_padding_duration': 0.0,
-                        'right_padding_duration': 0.0
-                    })
-                elif is_repaint_task or is_lego_task:
-                    # Repaint/lego task: May need padding for outpainting
-                    src_audio_duration = processed_src_audio.shape[-1] / 48000.0
-
-                    # Determine actual end time
-                    if repainting_end is None or repainting_end < 0:
-                        actual_end = src_audio_duration
-                    else:
-                        actual_end = repainting_end
-
-                    left_padding_duration = max(0, -repainting_start) if repainting_start is not None else 0
-                    right_padding_duration = max(0, actual_end - src_audio_duration)
-
-                    # Create padded audio
-                    left_padding_frames = int(left_padding_duration * 48000)
-                    right_padding_frames = int(right_padding_duration * 48000)
-
-                    if left_padding_frames > 0 or right_padding_frames > 0:
-                        # Pad the src audio
-                        batch_target_wavs = torch.nn.functional.pad(
-                            processed_src_audio,
-                            (left_padding_frames, right_padding_frames),
-                            'constant', 0
-                        )
-                    else:
-                        batch_target_wavs = processed_src_audio
-
-                    # Store padding info for coordinate adjustment
-                    padding_info_batch.append({
-                        'left_padding_duration': left_padding_duration,
-                        'right_padding_duration': right_padding_duration
-                    })
-                else:
-                    # Other tasks: Use src_audio directly without padding
-                    batch_target_wavs = processed_src_audio
-                    padding_info_batch.append({
-                        'left_padding_duration': 0.0,
-                        'right_padding_duration': 0.0
-                    })
-            else:
-                padding_info_batch.append({
-                    'left_padding_duration': 0.0,
-                    'right_padding_duration': 0.0
-                })
-                if audio_duration is not None and float(audio_duration) > 0:
-                    batch_target_wavs = self.create_target_wavs(float(audio_duration))
-                else:
-                    import random
-                    random_duration = random.uniform(10.0, 120.0)
-                    batch_target_wavs = self.create_target_wavs(random_duration)
-            target_wavs_batch.append(batch_target_wavs)
-
-        # Stack target_wavs into batch tensor
-        # Ensure all tensors have the same shape by padding to max length
-        max_frames = max(wav.shape[-1] for wav in target_wavs_batch)
-        padded_target_wavs = []
-        for wav in target_wavs_batch:
-            if wav.shape[-1] < max_frames:
-                pad_frames = max_frames - wav.shape[-1]
-                padded_wav = torch.nn.functional.pad(wav, (0, pad_frames), 'constant', 0)
-                padded_target_wavs.append(padded_wav)
-            else:
-                padded_target_wavs.append(wav)
-
-        target_wavs_tensor = torch.stack(padded_target_wavs, dim=0)  # [batch_size, 2, frames]
-
-        if can_use_repainting:
-            # Repaint task: Set repainting parameters
-            if repainting_start is None:
-                repainting_start_batch = None
-            elif isinstance(repainting_start, (int, float)):
-                if processed_src_audio is not None:
-                    adjusted_start = repainting_start + padding_info_batch[0]['left_padding_duration']
-                    repainting_start_batch = [adjusted_start] * actual_batch_size
-                else:
-                    repainting_start_batch = [repainting_start] * actual_batch_size
-            else:
-                # List input - adjust each item
-                repainting_start_batch = []
-                for i in range(actual_batch_size):
-                    if processed_src_audio is not None:
-                        adjusted_start = repainting_start[i] + padding_info_batch[i]['left_padding_duration']
-                        repainting_start_batch.append(adjusted_start)
-                    else:
-                        repainting_start_batch.append(repainting_start[i])
-
-            # Handle repainting_end - use src audio duration if not specified or negative
-            if processed_src_audio is not None:
-                # If src audio is provided, use its duration as default end
-                src_audio_duration = processed_src_audio.shape[-1] / 48000.0
-                if repainting_end is None or repainting_end < 0:
-                    # Use src audio duration (before padding), then adjust for padding
-                    adjusted_end = src_audio_duration + padding_info_batch[0]['left_padding_duration']
-                    repainting_end_batch = [adjusted_end] * actual_batch_size
-                else:
-                    # Adjust repainting_end to be relative to padded audio
-                    adjusted_end = repainting_end + padding_info_batch[0]['left_padding_duration']
-                    repainting_end_batch = [adjusted_end] * actual_batch_size
-            else:
-                # No src audio - repainting doesn't make sense without it
-                if repainting_end is None or repainting_end < 0:
-                    repainting_end_batch = None
-                elif isinstance(repainting_end, (int, float)):
-                    repainting_end_batch = [repainting_end] * actual_batch_size
-                else:
-                    # List input - adjust each item
-                    repainting_end_batch = []
-                    for i in range(actual_batch_size):
-                        if processed_src_audio is not None:
-                            adjusted_end = repainting_end[i] + padding_info_batch[i]['left_padding_duration']
-                            repainting_end_batch.append(adjusted_end)
-                        else:
-                            repainting_end_batch.append(repainting_end[i])
-        else:
-            # All other tasks (cover, text2music, extract, complete): No repainting
-            # Only repaint and lego tasks should have repainting parameters
-            repainting_start_batch = None
-            repainting_end_batch = None
-            
-        return repainting_start_batch, repainting_end_batch, target_wavs_tensor
-
     def _prepare_batch(
         self,
         captions: List[str],
